requirements.txt
ocrmypdf>=15.4.0
pathlib>=1.0.1
PyPDF2>=3.0.0
google-ai-generativelanguage>=0.4.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
langchain>=0.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
jupyter>=1.0.0
ipykernel>=6.0.0 